Working Version
This version is working , but it does not limit the maze to check positions that it already checked before, so it is probably running more times than it normally would need. As I am fucking lazy and it is just a for fun project I probably will not include that. It also won't return an error if there's no beginnin, ending or no solution and won't work if the number of columns is different in any row.